feat: add multi-value register
On concurrent writes, we will keep all values for which
we can't establish a causal history.

+++ tests/version_vector_test.cpp
@@ -8,7 +8,7 @@ auto main() -> int
 {
     assert(rc::check("increment method return the dot incremented by one",
             [](std::unordered_map<int, uint32_t> dots) {
-                version_vector<int> v(robin_hood::unordered_flat_map<int, uint32_t>(dots.begin(), dots.end()));
+                version_vector<int> v(robin_hood::unordered_flat_map<int, uint64_t>(dots.begin(), dots.end()));
                 for(auto [actor, counter] : v.dots) {
                     auto incremanted_dot = dot{actor, counter+1};
                     RC_ASSERT(v.inc(actor) == incremanted_dot);
